Organize Ui files
