+++ PalmHill.BlazorChat/Client/Components/Attachment/AttachmentItemDisplay.razor
