+++ PalmHill.BlazorChat/Client/Components/Attachment/FileInput.razor
@@ -1,8 +1,8 @@
-﻿@using Microsoft.AspNetCore.Components.Forms
-@using PalmHill.BlazorChat.ApiClient
-@using PalmHill.BlazorChat.Shared.Models
-@using Refit
-@inject BlazorChatApi BlazorChatApi
+﻿@using Microsoft.AspNetCore.Components.Forms    
+@using PalmHill.BlazorChat.ApiClient            
+@using PalmHill.BlazorChat.Shared.Models        
+@using Refit                                    
+@inject BlazorChatApi BlazorChatApi             
 
 <style>
     .file-upload-container {
