--- PalmHill.BlazorChat/Client/Components/SideNav.razor
@@ -1,5 +1,0 @@
-﻿<h3>SideNav</h3>
-
-@code {
-
-}
