+++ PalmHill.BlazorChat/Client/_Imports.razor
@@ -7,3 +7,9 @@
 @using PalmHill.BlazorChat.Client
 @using Microsoft.FluentUI.AspNetCore.Components
 @using PalmHill.BlazorChat.Client.Components;
+@using PalmHill.BlazorChat.Client.Components.Attachment;
+@using PalmHill.BlazorChat.Client.Components.Chat;
+@using PalmHill.BlazorChat.Client.Components.Markdown;
+@using PalmHill.BlazorChat.Client.Components.Settings;
+@using PalmHill.BlazorChat.Client.Models;
+@using PalmHill.BlazorChat.Shared.Models;
